Made CashSelectionTest a unit test by using mock network.
Also includes adding a default value of emptyList() to InternalMockNetwork.cordappPackages
