+++ core/src/test/kotlin/net/corda/core/flows/AttachmentTests.kt
@@ -30,7 +30,7 @@ class AttachmentTests {
 
     @Before
     fun setUp() {
-        mockNet = InternalMockNetwork(emptyList())
+        mockNet = InternalMockNetwork()
     }
 
     @After
