+++ node/src/test/kotlin/net/corda/node/internal/NetworkParametersTest.kt
@@ -29,8 +29,7 @@ import kotlin.test.assertFails
 
 class NetworkParametersTest {
     private val mockNet = InternalMockNetwork(
-            emptyList(),
-            MockNetworkParameters(networkSendManuallyPumped = true),
+            defaultParameters = MockNetworkParameters(networkSendManuallyPumped = true),
             notarySpecs = listOf(MockNetworkNotarySpec(DUMMY_NOTARY_NAME)))
 
     @After
