+++ node/src/test/kotlin/net/corda/node/messaging/InMemoryMessagingTests.kt
@@ -18,7 +18,7 @@ class InMemoryMessagingTests {
 
     @Before
     fun setUp() {
-        mockNet = InternalMockNetwork(emptyList())
+        mockNet = InternalMockNetwork()
     }
 
     @After
