+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/internal/InternalMockNetwork.kt
@@ -97,7 +97,7 @@ data class InternalMockNodeParameters(
     )
 }
 
-open class InternalMockNetwork(private val cordappPackages: List<String>,
+open class InternalMockNetwork(private val cordappPackages: List<String> = emptyList(),
                                defaultParameters: MockNetworkParameters = MockNetworkParameters(),
                                val networkSendManuallyPumped: Boolean = defaultParameters.networkSendManuallyPumped,
                                val threadPerNode: Boolean = defaultParameters.threadPerNode,
